# docs/static/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/signup": {
      "post": {
        "operationId": "AuthController_signup",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignupDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "userId": {
                      "type": "string",
                      "example": "user-1"
                    }
                  },
                  "required": ["userId"]
                }
              }
            }
          }
        },
        "summary": "Sign up user",
        "tags": ["Auth"],
        "description": "User signup. Tenant signup policy is enforced, passwords are hashed, duplicates are rejected, rate limiting applies, and signup activity is audited."
      }
    },
    "/auth/withdraw": {
      "post": {
        "operationId": "AuthController_withdraw",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WithdrawDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Withdraw current user",
        "tags": ["Auth"],
        "description": "Authenticated self-service account withdrawal. The caller must be authenticated; related sessions, tokens, consents, and audit records are handled by application policy."
      }
    },
    "/auth/password": {
      "put": {
        "operationId": "AuthController_changePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Change current user password",
        "tags": ["Auth"],
        "description": "Authenticated password change. Current password is verified, the new password must be different, password policy is enforced, and the change is audited."
      }
    },
    "/auth/password/reset-request": {
      "post": {
        "operationId": "AuthController_requestPasswordReset",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordResetRequestDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Request password reset",
        "tags": ["Auth"],
        "description": "Password reset request. Responses must avoid user enumeration, apply strong rate limiting, use short-lived one-time reset tokens, and emit security audit events."
      }
    },
    "/auth/password/reset": {
      "post": {
        "operationId": "AuthController_resetPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordResetDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Reset password",
        "tags": ["Auth"],
        "description": "Token-based password reset. Reset tokens are one-time and short-lived; successful reset revokes relevant sessions or tokens according to policy."
      }
    },
    "/auth/email/verification-request": {
      "post": {
        "operationId": "AuthController_requestEmailVerification",
        "parameters": [],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Request email verification",
        "tags": ["Auth"]
      }
    },
    "/auth/email/verify": {
      "post": {
        "operationId": "AuthController_verifyEmail",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerificationTokenDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Verify email",
        "tags": ["Auth"]
      }
    },
    "/auth/phone/verification-request": {
      "post": {
        "operationId": "AuthController_requestPhoneVerification",
        "parameters": [],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Request phone verification",
        "tags": ["Auth"]
      }
    },
    "/auth/phone/verify": {
      "post": {
        "operationId": "AuthController_verifyPhone",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerificationTokenDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Verify phone",
        "tags": ["Auth"]
      }
    },
    "/auth/mfa/totp/enroll": {
      "post": {
        "operationId": "AuthController_beginTotpEnrollment",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "secret": {
                      "type": "string",
                      "example": "JBSWY3DPEHPK3PXP"
                    },
                    "otpauthUrl": {
                      "type": "string",
                      "example": "otpauth://totp/Auth:john?secret=JBSWY3DPEHPK3PXP"
                    }
                  },
                  "required": ["secret", "otpauthUrl"]
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Begin TOTP enrollment",
        "tags": ["Auth"]
      }
    },
    "/auth/mfa/totp/confirm": {
      "post": {
        "operationId": "AuthController_confirmTotpEnrollment",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TotpConfirmationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "recoveryCodes": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "example": "ABCD-EFGH"
                      }
                    }
                  },
                  "required": ["recoveryCodes"]
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Confirm TOTP enrollment",
        "tags": ["Auth"]
      }
    },
    "/auth/mfa/totp": {
      "delete": {
        "operationId": "AuthController_disableTotp",
        "parameters": [],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Disable TOTP",
        "tags": ["Auth"]
      }
    },
    "/auth/mfa/recovery-codes/status": {
      "get": {
        "operationId": "AuthController_getRecoveryCodeStatus",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "remaining": {
                      "type": "integer",
                      "example": 8
                    },
                    "total": {
                      "type": "integer",
                      "example": 10
                    },
                    "used": {
                      "type": "integer",
                      "example": 2
                    },
                    "low": {
                      "type": "boolean",
                      "example": false
                    }
                  },
                  "required": ["remaining", "total", "used", "low"]
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get recovery code status",
        "tags": ["Auth"]
      }
    },
    "/auth/mfa/recovery-codes/rotate": {
      "post": {
        "operationId": "AuthController_rotateRecoveryCodes",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "recoveryCodes": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "example": "ABCD-EFGH"
                      }
                    }
                  },
                  "required": ["recoveryCodes"]
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Rotate recovery codes",
        "tags": ["Auth"]
      }
    },
    "/auth/mfa/preference": {
      "put": {
        "operationId": "AuthController_updateMfaPreference",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMfaPreferenceDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Update MFA preference",
        "tags": ["Auth"]
      }
    },
    "/auth/profile": {
      "get": {
        "operationId": "AuthController_getProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "example": "user-1"
                    },
                    "username": {
                      "type": "string",
                      "example": "john"
                    },
                    "email": {
                      "type": "string",
                      "example": "john@example.com",
                      "nullable": true
                    },
                    "emailVerified": {
                      "type": "boolean",
                      "example": false
                    },
                    "phone": {
                      "type": "string",
                      "example": "+821012345678",
                      "nullable": true
                    },
                    "phoneVerified": {
                      "type": "boolean",
                      "example": false
                    },
                    "status": {
                      "type": "string",
                      "example": "ACTIVE"
                    },
                    "mfaEnabled": {
                      "type": "boolean",
                      "example": false
                    },
                    "temporaryPassword": {
                      "type": "boolean",
                      "example": false
                    },
                    "passwordChangeRequired": {
                      "type": "boolean",
                      "example": false
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time",
                      "example": "2026-05-25T00:00:00.000Z"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time",
                      "example": "2026-05-25T00:00:00.000Z"
                    }
                  },
                  "required": [
                    "id",
                    "username",
                    "email",
                    "emailVerified",
                    "phone",
                    "phoneVerified",
                    "status",
                    "mfaEnabled",
                    "temporaryPassword",
                    "passwordChangeRequired",
                    "createdAt",
                    "updatedAt"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get current user profile",
        "tags": ["Auth"],
        "description": "Authenticated profile self-service. Password hashes, credential secrets, raw MFA material, and internal system fields are never exposed."
      },
      "put": {
        "operationId": "AuthController_updateProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Update current user profile",
        "tags": ["Auth"],
        "description": "Authenticated profile self-service. Password hashes, credential secrets, raw MFA material, and internal system fields are never exposed."
      }
    },
    "/auth/consents": {
      "get": {
        "operationId": "AuthController_getConsents",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "clientId": {
                        "type": "string",
                        "example": "web-app"
                      },
                      "clientName": {
                        "type": "string",
                        "example": "Web App"
                      },
                      "grantedScopes": {
                        "type": "string",
                        "example": "openid email profile"
                      },
                      "grantedAt": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-05-25T00:00:00.000Z"
                      }
                    },
                    "required": [
                      "clientId",
                      "clientName",
                      "grantedScopes",
                      "grantedAt"
                    ]
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "List current user consents",
        "tags": ["Auth"],
        "description": "Authenticated consent self-service. Users can list and revoke only their own client consents; revocation invalidates related refresh-token access according to policy."
      }
    },
    "/auth/identity-links": {
      "get": {
        "operationId": "AuthController_getIdentityLinks",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "example": "identity-1"
                      },
                      "provider": {
                        "type": "string",
                        "example": "google"
                      },
                      "email": {
                        "type": "string",
                        "example": "john@example.com",
                        "nullable": true
                      },
                      "linkedAt": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-05-25T00:00:00.000Z"
                      }
                    },
                    "required": ["id", "provider", "email", "linkedAt"]
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "List current user identity links",
        "tags": ["Auth"]
      }
    },
    "/auth/identity-links/{provider}/start": {
      "post": {
        "operationId": "AuthController_startIdentityLink",
        "parameters": [
          {
            "name": "provider",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartIdentityLinkDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "authorizationUrl": {
                      "type": "string",
                      "example": "https://idp.example.com/oauth/authorize?..."
                    }
                  },
                  "required": ["authorizationUrl"]
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Start identity provider link",
        "tags": ["Auth"]
      }
    },
    "/auth/identity-links/{provider}/callback": {
      "get": {
        "operationId": "AuthController_completeIdentityLink",
        "parameters": [
          {
            "name": "provider",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Redirect",
            "headers": {
              "Location": {
                "description": "Redirect target URL",
                "schema": {
                  "type": "string",
                  "example": "https://app.example.com/callback"
                }
              }
            }
          }
        },
        "summary": "Complete identity provider link",
        "tags": ["Auth"]
      }
    },
    "/auth/identity-links/{identityId}": {
      "delete": {
        "operationId": "AuthController_unlinkIdentity",
        "parameters": [
          {
            "name": "identityId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Unlink identity provider",
        "tags": ["Auth"]
      }
    },
...
  },
  "info": {
    "title": "Auth API",
    "description": "OIDC Authorization Server API Reference",
    "version": "0.1.1",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/",
      "description": "Service origin"
    },
    {
      "url": "/api",
      "description": "Vite proxy"
    }
  ],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "admin_session"
      }
    },
    "schemas": {
...
      "SignupDto": {
        "type": "object",
        "properties": {}
      },
      "WithdrawDto": {
        "type": "object",
        "properties": {}
      },
      "ChangePasswordDto": {
        "type": "object",
        "properties": {}
      },
      "PasswordResetRequestDto": {
        "type": "object",
        "properties": {}
      },
      "PasswordResetDto": {
        "type": "object",
        "properties": {}
      },
      "VerificationTokenDto": {
        "type": "object",
        "properties": {}
      },
      "TotpConfirmationDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateMfaPreferenceDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateProfileDto": {
        "type": "object",
        "properties": {}
      },
      "StartIdentityLinkDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}